\name{TempShift-package}
\alias{TempShift-package}
\alias{TempShift}
\docType{package}
\title{
TempShift: Identifying temporal shift of time series between multiple groups
}
\description{
TempShift identifies temporally-shifted genes through model comparison. It estimates the temporal shift between groups based on Gaussian process. 
}
\details{
\tabular{ll}{
Package: \tab TempShift\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-05-21\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Ying Zhu

Maintainer: Ying Zhu <zhuying04@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
